application/backend/application/requirements.txt
alembic==1.9.2
Flask==2.2.2
Flask-Cors==3.0.10
gunicorn==20.1.0
python-dotenv==1.0.0
psycopg2==2.9.5
SQLAlchemy==2.0.0
